Extract e2e_conftest.py as a real file instead of generating on the fly
